requirements.txt
langchain>=0.1.0
langchain-community>=0.0.27
langchain-openai>=0.0.5
openai>=1.14.0
chromadb>=0.4.22
tiktoken>=0.6.0
streamlit 

# Document processing
pypdf>=4.1.0
yt-dlp>=2024.3.10
pydub>=0.25.1
beautifulsoup4>=4.12.0
ffmpeg-python>=0.2.0

# Voice processing
SpeechRecognition>=3.10.0
openai-whisper>=20231117
pyaudio>=0.2.14

# GUI
panel>=1.3.1
param>=2.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.7 